Add functions for configuring DataCite (#46)
This change adds three functions for invoking the endpoints that get,
set, and test the DataCite configuration. I don't think that this is
particularly useful functionality for most API users as the
configuration is not expected to change often and thus wouldn't be
worthwhile automating, but it is useful for us where every AWS build or
re-build of localhost requires re-configuring the DataCite test
credentials.